requirements.txt
fastapi
uvicorn
face_recognition
pillow
python-multipart
